check if SDR healthcheck file exists before delete

+++ src/handlers/monitor_handler.py
@@ -1,13 +1,15 @@
 from sensor import settings
 from pathlib import Path
 import logging
+import os
 
 logger = logging.getLogger(__name__)
 
 def monitor_action_completed_callback(sender, **kwargs):
     healthy = kwargs["radio_healthy"]
     if healthy:
-        Path(settings.SDR_HEALTHCHECK_FILE).unlink()
+        if os.path.exists(settings.SDR_HEALTHCHECK_FILE):
+            Path(settings.SDR_HEALTHCHECK_FILE).unlink()
         logger.info("USRP healthy")
     else:
         logger.warning("USRP unhealthy")
